requirements.txt
Django==1.4.3
PIL==1.1.7
argparse==1.2.1
python-magic==0.4.3
wsgiref==0.1.2
